+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.2.0
+
+- Automatically download copy_tuner.yml
+- Add `copyTuner.download` command
+
 ## 0.1.0
 
 - Initial release